Version 1.0.3

+++ 2022-23_AM1_AM2_AM3_GanttManager_Public/src/main/java/sort/Compare.java
@@ -22,7 +22,7 @@ public class Compare {
 		for(Task i:topLevelTasks) {
 			int minStart = -1;
 			int maxEnd = 0;
-			int totalCost = 0;
+			double totalCost = 0;
 			boolean hasSubtask = false;
 			ArrayList<Task> subtaskList = new ArrayList<Task>();
 			for (Task j:taskList) {
